Merged in ejether/fixing-definitions-reference (pull request #1)
Ejether/fixing definitions reference

* Fix Definitions Reference

There was an issue where the namespace definitions was misbehaving. It appears the culprit was the `helpers.merge_dict` function that was modifying the definitions dictionary instead of returning a copy. I updated the code to return a copy instead and monitors got generated as inteneded then.

* expanded TerraDog.Resource tests

* missed resource_name var